requirements_pro.txt
# AI Vision Framework Pro - Enhanced Requirements

# Core Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Numerical Computing
numpy>=1.24.0
scipy>=1.11.0

# Image Processing
pillow>=10.0.0
scikit-image>=0.21.0
albumentations>=1.3.0

# Object Detection
ultralytics>=8.0.0

# Model Hub & Advanced Models
huggingface-hub>=0.16.0
transformers>=4.30.0
timm>=0.9.0

# Advanced Models (Optional but Recommended)
# CLIP for zero-shot classification
# git+https://github.com/openai/CLIP.git

# Model Formats
onnx>=1.14.0
onnxruntime>=1.15.0

# Face Recognition
face-recognition>=1.3.0
mtcnn>=0.1.1
dlib>=19.24.0

# Gesture Recognition
mediapipe>=0.10.0

# REST API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Handling
pandas>=2.0.0
pyyaml>=6.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0

# Video Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Object Tracking (Optional)
# filterpy>=1.4.5  # For Kalman filters
# lap>=0.4.0  # For Hungarian algorithm

# Database (Optional for Face Recognition)
# pymongo>=4.4.0  # For MongoDB
# firebase-admin>=6.2.0  # For Firebase

# Cloud Integration (Optional)
# google-cloud-vision>=3.4.0  # Google Cloud Vision API
# boto3>=1.28.0  # AWS SDK for Rekognition

# Security
cryptography>=41.0.0

# Development Tools (Optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0

# GUI Enhancements
# Note: tkinter comes with Python, but on some Linux systems:
# sudo apt-get install python3-tk

# Performance Monitoring
psutil>=5.9.0
GPUtil>=1.4.0  # For GPU monitoring
